Allow passing a callback_url param
When building an application we need to do two types of Twitter
authentication - one for signing in to our application (and creating a
`User`) and another for associating Twitter accounts to their `User`
account.

By passing a custom `callback_url` param we can have one controller
method for signing in and another one for the associating part and just
have the application set the callback when starting the authentication
process.